+++ rye/src/cli/init.rs
@@ -23,7 +23,7 @@ use crate::platform::{
     get_python_version_request_from_pyenv_pin,
 };
 use crate::pyproject::BuildSystem;
-use crate::sources::PythonVersionRequest;
+use crate::sources::py::PythonVersionRequest;
 use crate::utils::{
     copy_dir, escape_string, format_requirement, get_venv_python_bin, is_inside_git_work_tree,
     CommandOutput, CopyDirOptions, IoPathContext,
@@ -688,7 +688,7 @@ fn get_setup_py_json<T: AsRef<Path>>(path: T, python: T) -> Result<Value, Error>
 ///  * `<archive_url>`. TBD
 ///  * Options without a requirement, such as `--find-links` or `--index-url`
 ///
-/// See https://github.com/mitsuhiko/rye/issues/191
+/// See https://github.com/astral-sh/rye/issues/191
 fn import_requirements_file(
     requirements: &mut BTreeMap<String, String>,
     path: impl AsRef<Path>,