+++ rye/src/cli/rye.rs
@@ -21,7 +21,7 @@ use crate::bootstrap::{
 use crate::cli::toolchain::register_toolchain;
 use crate::config::Config;
 use crate::platform::{get_app_dir, symlinks_supported};
-use crate::sources::{get_download_url, PythonVersionRequest};
+use crate::sources::py::{get_download_url, PythonVersionRequest};
 use crate::utils::{check_checksum, toml, tui_theme, CommandOutput, IoPathContext, QuietExit};
 
 #[cfg(windows)]
@@ -29,7 +29,7 @@ const DEFAULT_HOME: &str = "%USERPROFILE%\\.rye";
 #[cfg(unix)]
 const DEFAULT_HOME: &str = "$HOME/.rye";
 
-const GITHUB_REPO: &str = "https://github.com/mitsuhiko/rye";
+const GITHUB_REPO: &str = "https://github.com/astral-sh/rye";
 const UNIX_ENV_FILE: &str = r#"
 # rye shell setup
 {%- if custom_home %}
@@ -197,7 +197,7 @@ fn update(args: UpdateCommand) -> Result<(), Error> {
         let tmp = tempdir()?;
         cmd.arg("install")
             .arg("--git")
-            .arg("https://github.com/mitsuhiko/rye")
+            .arg("https://github.com/astral-sh/rye")
             .arg("--root")
             .env(
                 "PATH",
@@ -344,6 +344,7 @@ fn uninstall(args: UninstallCommand) -> Result<(), Error> {
         remove_dir_all_if_exists(&app_dir.join("self"))?;
         remove_dir_all_if_exists(&app_dir.join("py"))?;
         remove_dir_all_if_exists(&app_dir.join("pip-tools"))?;
+        remove_dir_all_if_exists(&app_dir.join("uv"))?;
         remove_dir_all_if_exists(&app_dir.join("tools"))?;
 
         // special deleting logic if we are placed in the app dir and the shim deletion