+++ rye/tests/test_sync.rs
@@ -1,3 +1,5 @@
+use std::fs;
+
 use crate::common::{rye_cmd_snapshot, Space};
 
 mod common;
@@ -11,23 +13,31 @@ fn test_empty_sync() {
     exit_code: 0
     ----- stdout -----
     Initializing new virtualenv in [TEMP_PATH]/project/.venv
-    Python version: cpython@3.12.1
+    Python version: cpython@3.12.2
     Generating production lockfile: [TEMP_PATH]/project/requirements.lock
     Generating dev lockfile: [TEMP_PATH]/project/requirements-dev.lock
     Installing dependencies
     Done!
 
     ----- stderr -----
-    warning: Requirements file [TEMP_FILE] does not contain any dependencies
-    Built 1 editable in [EXECUTION_TIME]
-    Resolved 1 package in [EXECUTION_TIME]
-    warning: Requirements file [TEMP_FILE] does not contain any dependencies
-    Built 1 editable in [EXECUTION_TIME]
-    Resolved 1 package in [EXECUTION_TIME]
     Built 1 editable in [EXECUTION_TIME]
     Installed 1 package in [EXECUTION_TIME]
      + my-project==0.1.0 (from file:[TEMP_PATH]/project)
     "###);
+
+    // is the prompt set?
+    #[cfg(unix)]
+    {
+        let script = space.venv_path().join("bin/activate");
+        let contents = fs::read_to_string(script).unwrap();
+        assert!(contents.contains("VIRTUAL_ENV_PROMPT=\"my-project\""));
+    }
+    #[cfg(windows)]
+    {
+        let script = space.venv_path().join("Scripts/activate.bat");
+        let contents = fs::read_to_string(script).unwrap();
+        assert!(contents.contains("@set \"VIRTUAL_ENV_PROMPT=my-project\""));
+    }
 }
 
 #[test]
@@ -41,9 +51,9 @@ fn test_add_and_sync_no_auto_sync() {
     exit_code: 0
     ----- stdout -----
     Initializing new virtualenv in [TEMP_PATH]/project/.venv
-    Python version: cpython@3.12.1
-    Added colorama>=0.4.6 as regular dependency
+    Python version: cpython@3.12.2
     Added flask>=3.0.0 as regular dependency
+    Added colorama>=0.4.6 as regular dependency
 
     ----- stderr -----
     "###);
@@ -58,12 +68,6 @@ fn test_add_and_sync_no_auto_sync() {
     Done!
 
     ----- stderr -----
-    warning: Requirements file [TEMP_FILE] does not contain any dependencies
-    Built 1 editable in [EXECUTION_TIME]
-    Resolved 9 packages in [EXECUTION_TIME]
-    warning: Requirements file [TEMP_FILE] does not contain any dependencies
-    Built 1 editable in [EXECUTION_TIME]
-    Resolved 9 packages in [EXECUTION_TIME]
     Built 1 editable in [EXECUTION_TIME]
     Resolved 8 packages in [EXECUTION_TIME]
     Downloaded 8 packages in [EXECUTION_TIME]
@@ -90,9 +94,9 @@ fn test_add_autosync() {
     exit_code: 0
     ----- stdout -----
     Initializing new virtualenv in [TEMP_PATH]/project/.venv
-    Python version: cpython@3.12.1
-    Added colorama>=0.4.6 as regular dependency
+    Python version: cpython@3.12.2
     Added flask>=3.0.0 as regular dependency
+    Added colorama>=0.4.6 as regular dependency
     Reusing already existing virtualenv
     Generating production lockfile: [TEMP_PATH]/project/requirements.lock
     Generating dev lockfile: [TEMP_PATH]/project/requirements-dev.lock
@@ -100,12 +104,6 @@ fn test_add_autosync() {
     Done!
 
     ----- stderr -----
-    warning: Requirements file [TEMP_FILE] does not contain any dependencies
-    Built 1 editable in [EXECUTION_TIME]
-    Resolved 9 packages in [EXECUTION_TIME]
-    warning: Requirements file [TEMP_FILE] does not contain any dependencies
-    Built 1 editable in [EXECUTION_TIME]
-    Resolved 9 packages in [EXECUTION_TIME]
     Built 1 editable in [EXECUTION_TIME]
     Resolved 8 packages in [EXECUTION_TIME]
     Downloaded 8 packages in [EXECUTION_TIME]