+++ docs/changelog.md
@@ -2,7 +2,7 @@
 
 Here you can find all the released changes to Rye.  If you want to also see
 the in-development changes that were not released yet, refer to the
-[CHANGELOG.md](https://github.com/mitsuhiko/rye/blob/main/CHANGELOG.md) file
+[CHANGELOG.md](https://github.com/astral-sh/rye/blob/main/CHANGELOG.md) file
 in the repository.
 
 {%